Continued Vampiric Grasp 1.1
Removed colliders on the vampiric claw prefab as recommended by Eric. Solved the issue of pushing the player upward but now idk how to make the prefab instantiate protruding from the player character since its also a prefab and I don't want to break it just to add my prefab.