Fireball Fix
Fixed the problems mentioned in the previous fireball commit.  The fireball had an extra rigidbody component that it didn't need, so I removed that.  I deleted the fixedupdate portion of the code because it wasn't necessary.  I calculated the starting position as well as the necessary speed, and got the fireball to move.  The spell is still not complete because you can spam fireballs into existence, also only one fireball currently works at a time because all of the clones are named the same.  Getting multiple fireballs to work as expected will take a few more lines of code.

Additionally I fixed bugs with the new movement script. The new movement script accidentally broke the camera movement and realm switching, which now work.